requirements.txt
numpy
yfinance
matplotlib.pyplot
scipy.stats